fix(AST): set correct Rust type for Java's `long`

+++ javac-rs-ast/src/ast.rs
@@ -9,7 +9,7 @@ type Char = u16;
 /// Java `int` type
 type Int = i32;
 /// Java `long` type
-type Long = i32;
+type Long = i64;
 /// Java `float` type
 type Float = f32;
 /// Java `double` type
